Update Solr schema.
We were removing stopwords from *all* indexed versions of the full
document text, which is a big problem when you want to do proper
text analysis.  We now have a complete stored (though lowercased
and w/o punctuation) version of the plain text, from which stop
words are only removed on *query*.  This means that in the index,
and hence in the term vectors, we will get information about the
stop words.

This also precipitated a general schema cleanup; the various text
field types are now strict supersets of one another, which should
make searching behavior easier to debug and maintain.