Migrate from delayed_job to Resque (closes #44).
This commit is, unfortunately, *way* too big.  I *guarantee* there
are bugs in here.  We've simultaneously had to switch from DJ to RQ,
move file attachments into the database using Paperclip and the
paperclip_database storage backend, deploy a new controller action
for showing uploaded asset images, update a bunch of our job parameter
passing logic, rework the specs, rework Airbrake error handling, and
so on, and so on.

The specs pass, but I'm going to start now to bash on some testing
of the job workers, to see if things actually work as they're
supposed to.