Refactor analysis tasks as a dataset nested resource.
This is a massive change, including the introduction of a new
AnalysisTasksController, the refactoring of AnalysisTask to
Datasets::AnalysisTask, and reworking the flow through the
creation of analysis tasks (formerly the datasets#task_start
view).  More refactoring yet needs to be done (especially in
datasets/analysis_tasks#new), but this is a great start.