Remove the ability to manually start analysis tasks.
This saves us from the silly hacks concerning adding the right
number of datasets to multiple-dataset analysis tasks.  Now the
only way to start tasks is by using the workflow controller,
which is the plausible way to do this anyway, and the one that
will actually make sense to users.