Move translation of languages out of helper into core_ext.
This should really be patched in as a method on
I18n.available_locales, so just go ahead and do it the right
way. This may wind up being brittle, but I think it's technically
the right place for that code, as it really is an accessor on
the list of available locales.